gold standard run for patients

+++ 1.max_projection/scripts/0.cp_max_projection.py
@@ -5,62 +5,48 @@
 
 # ## Import libraries
 
-# In[ ]:
+# In[1]:
 
 
-import argparse
+import os
 import pathlib
-import sys
 
 import tifffile
 import tqdm
+from arg_parsing_utils import check_for_missing_args, parse_args
+from notebook_init_utils import bandicoot_check, init_notebook
 
-# check if in a jupyter notebook
-try:
-    cfg = get_ipython().config
-    in_notebook = True
-except NameError:
-    in_notebook = False
+root_dir, in_notebook = init_notebook()
+image_base_dir = bandicoot_check(
+    pathlib.Path(os.path.expanduser("~/mnt/bandicoot")).resolve(), root_dir
+)
 
-# Get the current working directory
-cwd = pathlib.Path.cwd()
-
-if (cwd / ".git").is_dir():
-    root_dir = cwd
-
-else:
-    root_dir = None
-    for parent in cwd.parents:
-        if (parent / ".git").is_dir():
-            root_dir = parent
-            break
-
-# Check if a Git root directory was found
-if root_dir is None:
-    raise FileNotFoundError("No Git root directory found.")
-
-sys.path.append(f"{root_dir}/utils/")
-
-from parsable_args import parse_featurization_args_patient
 
 # In[2]:
 
 
-
 if not in_notebook:
-    args_dict = parse_featurization_args_patient()
+    args_dict = parse_args()
     patient = args_dict["patient"]
+    check_for_missing_args(
+        patient=patient,
+    )
 else:
-    patient = "NF0014"
+    patient = "NF0016_T1"
 
 
 # In[3]:
 
 
 # input images directory
-images_dir = pathlib.Path(f"../../data/{patient}/zstack_images/").resolve(strict=True)
+images_dir = pathlib.Path(f"{image_base_dir}/data/{patient}/zstack_images/").resolve(
+    strict=True
+)
 # output images directory
-output_dir = pathlib.Path(f"../../data/{patient}/zmax_proj/").resolve()
+output_dir = pathlib.Path(
+    f"{image_base_dir}/data/{patient}/2D_analysis/0a.zmax_proj/"
+).resolve()
+output_dir.mkdir(parents=True, exist_ok=True)
 
 
 # In[4]:
@@ -70,11 +56,14 @@ output_dir = pathlib.Path(f"../../data/{patient}/zmax_proj/").resolve()
 tiff_files = list(images_dir.rglob("*.tif"))
 tiff_files.sort()
 for tiff_file in tqdm.tqdm(tiff_files):
+    output_file_dir = output_dir / str(tiff_file.parent).split("/")[-1] / tiff_file.name
+    output_file_dir.parent.mkdir(parents=True, exist_ok=True)
+    if output_file_dir.exists():
+        continue
     # load the first tiff file to get the metadata
     image = tifffile.TiffFile(tiff_file)
     # z max projection
     max_proj = image.asarray().max(axis=0)
     # save the middle slice as a new tiff file
-    output_file_dir = output_dir / str(tiff_file.parent).split("/")[-1] / tiff_file.name
-    output_file_dir.parent.mkdir(parents=True, exist_ok=True)
+
     tifffile.imwrite(output_file_dir, max_proj)
